+++ ci/typescript-test.ts
@@ -4,4 +4,5 @@
  * Our index.d.ts is manually maintained.
  * This import ensures that the types are compiling with no errors.
  */
+// oxlint-disable-next-line no-unused-vars
 import * as snowflake from '../dist';